DownloadFileAsync/ExtractZipAsyncをProgressiveIoに集約しtempFile残存を修正
