+++ YukkuriMovieMaker.Plugin.Community/Voice/PiperPlus/Resource/PiperBinaryResource.cs
@@ -1,6 +1,4 @@
 using System.IO;
-using System.IO.Compression;
-using System.Net.Http;
 using System.Runtime.InteropServices;
 using YukkuriMovieMaker.Commons;
 
@@ -166,92 +164,17 @@ internal static class PiperBinaryResource
     {
         var zipPath = Path.Combine(tempDir, assetName);
         var downloadUrl = $"https://github.com/{RepoOwner}/{RepoName}/releases/download/{version}/{assetName}";
+        var client = HttpClientFactory.Client;
 
-        await DownloadFileAsync(downloadUrl, zipPath,
+        await ProgressiveIo.DownloadFileAsync(client, downloadUrl, zipPath,
             startFraction: 0.0, endFraction: 0.9,
             Texts.DownloadingBinary, progress, cancellationToken);
 
-        await ExtractZipAsync(zipPath, tempDir,
+        await ProgressiveIo.ExtractZipAsync(zipPath, tempDir,
             startFraction: 0.9, endFraction: 1.0,
             Texts.ExtractingBinary, progress, cancellationToken);
 
         File.Delete(zipPath);
-    }
-
-    static async Task DownloadFileAsync(
-        string url,
-        string destinationPath,
-        double startFraction,
-        double endFraction,
-        string message,
-        IProgress<(double Progress, string Message)>? progress,
-        CancellationToken cancellationToken)
-    {
-        var client = HttpClientFactory.Client;
-        using var response = await client.GetAsync(url, HttpCompletionOption.ResponseHeadersRead, cancellationToken);
-        response.EnsureSuccessStatusCode();
-
-        var totalBytes = response.Content.Headers.ContentLength ?? -1L;
-
-        await using var contentStream = await response.Content.ReadAsStreamAsync(cancellationToken);
-        await using var fileStream = File.Create(destinationPath);
-
-        var buffer = new byte[81920];
-        long downloaded = 0;
-        int read;
-        while ((read = await contentStream.ReadAsync(buffer, cancellationToken)) > 0)
-        {
-            await fileStream.WriteAsync(buffer.AsMemory(0, read), cancellationToken);
-            downloaded += read;
-            if (totalBytes > 0)
-            {
-                var fileFraction = (double)downloaded / totalBytes;
-                var overall = startFraction + fileFraction * (endFraction - startFraction);
-                progress?.Report((overall, message));
-            }
-        }
-    }
-
-    static async Task ExtractZipAsync(
-        string zipPath,
-        string destinationDir,
-        double startFraction,
-        double endFraction,
-        string message,
-        IProgress<(double Progress, string Message)>? progress,
-        CancellationToken cancellationToken)
-    {
-        await Task.Run(() =>
-        {
-            using var archive = ZipFile.OpenRead(zipPath);
-            var entries = archive.Entries;
-            var total = entries.Count;
-
-            for (var i = 0; i < total; i++)
-            {
-                cancellationToken.ThrowIfCancellationRequested();
-
-                var entry = entries[i];
-                var destPath = Path.GetFullPath(Path.Combine(destinationDir, entry.FullName));
-
-                if (!destPath.StartsWith(Path.GetFullPath(destinationDir) + Path.DirectorySeparatorChar, StringComparison.Ordinal))
-                    throw new InvalidOperationException($"Entry '{entry.FullName}' would extract outside destination directory.");
-
-                if (entry.FullName.EndsWith('/') || entry.FullName.EndsWith('\\'))
-                {
-                    Directory.CreateDirectory(destPath);
-                }
-                else
-                {
-                    Directory.CreateDirectory(Path.GetDirectoryName(destPath)!);
-                    entry.ExtractToFile(destPath, overwrite: true);
-                }
-
-                var fraction = (double)(i + 1) / total;
-                var overall = startFraction + fraction * (endFraction - startFraction);
-                progress?.Report((overall, message));
-            }
-        }, cancellationToken);
     }
 
     static void CommitInstall(
